Analysis modules separated into header and cpp files
