+++ src/modules/SalientFreqs.h
@@ -18,94 +18,25 @@
 // SalientFreqs inherits from the ModuleInterface with an int* output type
 class SalientFreqs : public ModuleInterface<int*> {
 public:
-    int numFreqs;
-    int* salientFreqs;
-    float* amplitudes;
+    int             numFreqs;
+    int*            salientFreqs;
+    float*          amplitudes;
     DeltaAmplitudes deltaAmps = DeltaAmplitudes();
 
-    SalientFreqs()
-    {
-        numFreqs = 3; // default to finding frequency of max change
-        salientFreqs = new int[numFreqs];
-        this->addSubmodule(&deltaAmps);
+    SalientFreqs();
 
-        for (int i = 0; i < numFreqs; i++) {
-            salientFreqs[i] = -1;
-        }
-    }
+    SalientFreqs(int n);
 
-    SalientFreqs(int n)
-    {
-        if (n < 0) {
-            n = 0;
-        }
-        if (n > windowSizeBy2) {
-            n = windowSizeBy2;
-        }
-
-        numFreqs = n;
-        salientFreqs = new int[numFreqs];
-        this->addSubmodule(&deltaAmps);
-
-        for (int i = 0; i < numFreqs; i++) {
-            salientFreqs[i] = -1;
-        }
-    }
-
-    ~SalientFreqs()
-    {
-        delete[] salientFreqs;
-    }
+    ~SalientFreqs();
 
     // change the amount of salient frequencies to be found
-    void changeNumFreqs(int newSize)
-    {
-        numFreqs = newSize;
-        delete[] salientFreqs;
-        salientFreqs = new int[numFreqs];
-    }
+    void changeNumFreqs(int newSize);
 
     // finds the n (numFreqs) bins with highest change in amplitude, stored in salientFreqs[]
     // input is assumed to be a 2D array of FFT data passed in from the Vibrosonics class
-    void doAnalysis()
-    {
-        deltaAmps.doAnalysis();
-        amplitudes = deltaAmps.getOutput(); // copy amplitudes
+    void doAnalysis();
 
-        // iterate through amplitudes to find the maximum(s)
-        int currMaxAmp = 0;
-        int currMaxAmpIdx = -1;
-        for (int i = 0; i < numFreqs; i++) {
-            for (int j = lowerBinBound; j < upperBinBound; j++) {
-                if (amplitudes[j] > currMaxAmp) {
-                    currMaxAmp = amplitudes[j];
-                    currMaxAmpIdx = j;
-                }
-            }
-            salientFreqs[i] = currMaxAmpIdx; // add the new max amp index to the array
-            amplitudes[currMaxAmpIdx] = 0; // set the amp to 0 so it will not be found again
-            currMaxAmp = 0; // reset iterators
-            currMaxAmpIdx = -1;
-        }
-        output = salientFreqs;
-
-        // if debug is enabled, print the output to the serial console
-        if (debugMode & DEBUG_ENABLE) {
-            Serial.printf("===SALIENT_FREQS===\n");
-            if (debugMode & DEBUG_VERBOSE) {
-                printModuleInfo();
-            }
-            Serial.printf("Salient Freqs: ");
-            for (int i = 0; i < numFreqs; i++) {
-                Serial.printf("%d ", salientFreqs[i]);
-            }
-            Serial.printf("\n==========================\n");
-        }
-    }
-
-    void printOutput()
-    {
-    }
+    void printOutput();
 };
 
 #endif
